Separate running jobs from failed jobs
